refactor: Update useFavoritePodcasts to load favorites from CloudStorage
This commit updates the `useFavoritePodcasts` module to load favorites from the `CloudStorage` instead of using a commented out code block. It removes the commented out code and replaces it with a call to `window.Telegram.WebApp.CloudStorage.getItem` to retrieve the favorites. If there is an error, it shows an alert with the error message. If there is a value, it sets the favorites using `JSON.parse(value)`.