refactor: Update SearchPage and Favorite components
This commit updates the SearchPage component to add the autoFocus prop to the Search component, allowing the search input to be automatically focused when the page loads. It also updates the Favorite component to add a max-width class to the Button component, limiting its width to 60% of the parent container. These changes improve the usability and visual appearance of the components.