refactor: Update page component to include Search and ContinueListening containers
This commit updates the page component to include the Search and ContinueListening containers. The Search container is added to display a search input field, allowing users to search for podcasts. The ContinueListening container is added to display a list of episodes that the user has previously listened to. These changes enhance the functionality and user experience of the application.